split set factory
